broke down ldap module
