+++ cas-server-support-ldap-service-registry/.gitignore
@@ -1,0 +1,2 @@
+ad.properties
+openldap.properties
