+++ cas-server-support-ldap-service-registry/src/test/resources/ldap-regservice-test.xml
