+++ cas-server-support-ldap-service-registry/src/test/resources/services/sampleService.json
@@ -1,0 +1,28 @@
+{
+    "@class" : "org.jasig.cas.services.RegexRegisteredService",
+    "serviceId" : "^(https|imaps)://.*",
+    "name" : "HTTPS and IMAPS",
+    "id" : 10000001,
+    "description" : "This service definition authorized all application urls that support HTTPS and IMAPS protocols.",
+    "proxyPolicy" : {
+        "@class" : "org.jasig.cas.services.RefuseRegisteredServiceProxyPolicy"
+    },
+    "evaluationOrder" : 10000,
+    "usernameAttributeProvider" : {
+        "@class" : "org.jasig.cas.services.DefaultRegisteredServiceUsernameProvider"
+    },
+    "logoutType" : "BACK_CHANNEL",
+    "attributeReleasePolicy" : {
+        "@class" : "org.jasig.cas.services.ReturnAllowedAttributeReleasePolicy",
+        "principalAttributesRepository" : {
+            "@class" : "org.jasig.cas.authentication.principal.DefaultPrincipalAttributesRepository"
+        },
+        "authorizedToReleaseCredentialPassword" : false,
+        "authorizedToReleaseProxyGrantingTicket" : false
+    },
+    "accessStrategy" : {
+        "@class" : "org.jasig.cas.services.DefaultRegisteredServiceAccessStrategy",
+        "enabled" : true,
+        "ssoEnabled" : true
+    }
+}
