+++ cas-server-support-ldap/pom.xml
@@ -25,16 +25,9 @@
         <version>4.2.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-  <artifactId>cas-server-support-ldap</artifactId>
+    <artifactId>cas-server-support-ldap</artifactId>
     <packaging>jar</packaging>
     <name>Apereo CAS LDAP Support</name>
-
-  <repositories>
-    <repository>
-      <id>vt-middleware</id>
-      <url>https://raw.github.com/vt-middleware/maven-repo/master</url>
-    </repository>
-  </repositories>
 
   <developers>
     <developer>
