--- cas-server-support-ldap/src/main/resources/cas.properties.example
@@ -1,65 +1,0 @@
-#========================================
-# General properties
-#========================================
-ldap.url=ldap://localhost:1389
-
-# Start TLS for SSL connections
-ldap.useStartTLS=false
-
-# Directory root DN
-ldap.rootDn=dc=example,dc=org
-
-# Base DN of users to be authenticated
-ldap.baseDn=ou=people,dc=example,dc=org
-
-# LDAP connection timeout in milliseconds
-ldap.connectTimeout=3000
-
-# Manager credential DN
-ldap.managerDn=cn=Directory Manager,dc=example,dc=org
-
-# Manager credential password
-ldap.managerPassword=Password
-
-#========================================
-# LDAP connection pool configuration
-#========================================
-ldap.pool.minSize=1
-ldap.pool.maxSize=10
-ldap.pool.validateOnCheckout=false
-ldap.pool.validatePeriodically=true
-
-# Amount of time in milliseconds to block on pool exhausted condition
-# before giving up.
-ldap.pool.blockWaitTime=3000
-
-# Frequency of connection validation in seconds
-# Only applies if validatePeriodically=true
-ldap.pool.validatePeriod=300
-
-# Attempt to prune connections every N seconds
-ldap.pool.prunePeriod=300
-
-# Maximum amount of time an idle connection is allowed to be in
-# pool before it is liable to be removed/destroyed
-ldap.pool.idleTime=600
-
-#========================================
-# Authentication
-#========================================
-ldap.authn.searchFilter=cn={user}
-
-# Ldap domain used to resolve dn
-ldap.domain=example.org
-
-#========================================
-# User Details
-#========================================
-ldap.role.baseDn=ou=people,dc=example,dc=org
-ldap.role.searchFilter=&(objectclass=casServiceUserDetails)(member={0})
-ldap.user.searchFilter=&(objectclass=casServiceUserDetails)(mail={0})
-
-#========================================
-# Registered Service
-#========================================
-ldap.service.baseDn=ou=people,dc=example,dc=org
