+++ cas-server-support-ldap/src/test/java/org/jasig/cas/AllTestsSuite.java
@@ -18,7 +18,6 @@
  */
 package org.jasig.cas;
 
-import org.jasig.cas.adaptors.ldap.services.LdapServiceRegistryDaoTests;
 import org.jasig.cas.authentication.LdapAuthenticationHandlerTests;
 import org.jasig.cas.monitor.ConnectionFactoryMonitorTests;
 import org.jasig.cas.monitor.PooledConnectionFactoryMonitorTests;
@@ -33,7 +32,6 @@ import org.junit.runners.Suite;
  */
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
-    LdapServiceRegistryDaoTests.class,
     LdapAuthenticationHandlerTests.class,
     ConnectionFactoryMonitorTests.class,
     PooledConnectionFactoryMonitorTests.class,
