+++ pom.xml
@@ -1405,6 +1405,7 @@
         <module>cas-server-support-jpa-ticket-registry</module>
         <module>cas-server-support-jpa-service-registry</module>
         <module>cas-server-support-ldap</module>
+        <module>cas-server-support-ldap-service-registry</module>
         <module>cas-server-support-openid</module>
         <module>cas-server-support-wsfederation</module>
         <module>cas-server-support-radius</module>
